requirements.txt
discord.py==2.3.2
atproto==0.0.65
python-dotenv==1.0.0
